requirements.txt
Django==4.2.7
python-pptx==0.6.21
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0 